refactor: 승부의 판정을 딜러로 이동한다
