+++ src/main/kotlin/blackjack/domain/game/BlackJackGame.kt
@@ -1,21 +1,19 @@
 package blackjack.domain.game
 
 import blackjack.domain.card.Deck
+import blackjack.domain.player.Player
 import blackjack.domain.player.Players
 import blackjack.views.OutputView
 
-class BlackJackGame(private val players: Players, private val deck: Deck) {
+class BlackJackGame(var players: Players, private val deck: Deck) {
 
     fun play(
         isPlayerTurnOff: Turn,
-        initCallback: (Players) -> Unit,
-        playingCallback: (Players) -> Unit,
-    ): Players {
+        initCallback: (Players) -> Unit
+    ) {
         val afterInitPhased = players.startInitPhase(deck)
         initCallback(afterInitPhased)
-        val playingPhasedPlayers = playingPhase(isPlayerTurnOff, deck, afterInitPhased)
-        playingCallback(playingPhasedPlayers)
-        return playingPhasedPlayers
+        this.players = playingPhase(isPlayerTurnOff, deck, afterInitPhased)
     }
 
     private fun playingPhase(isPlayerTurnOff: Turn, deck: Deck, players: Players): Players {
@@ -30,5 +28,9 @@ class BlackJackGame(private val players: Players, private val deck: Deck) {
 
     private fun receiveCardAllPlayers(isPlayerTurnOff: Turn, deck: Deck, players: Players): Players {
         return players.receiveCards(deck, isPlayerTurnOff, OutputView::printCards, OutputView::printDealerCardReceived)
+    }
+
+    fun getResult(): Map<Player, List<Score>> {
+        return players.getResult()
     }
 }
