+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -2,7 +2,6 @@ package blackjack.domain.player
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.game.Rule
 import blackjack.domain.game.Score
 
 interface Player {
@@ -22,5 +21,5 @@ interface Player {
 
     fun canReceiveCard(): Boolean
 
-    fun judgeResult(players: List<Player>, rule: Rule): Map<Player, List<Score>>
+    fun judge(players: List<Player>): Map<Player, List<Score>>
 }
